Changes according to daniela's suggestions:
"I just looked over the changes and there are 2 small additional edits:

1) Please remove this text which is at the end of the discussion section because it talks about manipulators

An alternative ap- proach to robot manipulator design is to allow soft segments to handle some uncertainty at the hardware level in order to reduce the burden on the computational system. For exam- ple, consider how a soft manipulator passively conforms to the environment’s boundary. A planner is unaware of this com- plex interaction, but the primary task can still be successfully executed. Further, modern inspection tasks as well as invasive surgery require devices with redundant degrees of freedom and high dexterity and often impose the constraint of navigating around sensitive objects. Soft robotic manipulators may be well-suited for this class of tasks.

2) In the intro, the last paragraph should be This paper is organized as follows... etc Please remove the contributions (that start with Specifically this paper contributes...) above this text

daniela"